requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langgraph
langsmith
tavily-python
groq
streamlit
faiss-cpu
